% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaxsubmit.R
\name{clean.remote}
\alias{clean.remote}
\title{Clean .jaxsubmit directory}
\usage{
clean.remote(target.dir = "~/")
}
\arguments{
\item{target.dir}{Path to the parent directory of .jaxsubmit/}
}
\description{
This function is a helper function to delete the .jaxsubmit directory created when the
jaxsubmit::submit() command is run.
}
