% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{JobProfile-class}
\alias{JobProfile-class}
\alias{JobProfile}
\title{JobProfile constructor}
\description{
An S4 class used to create a job specification.
}
\section{Slots}{

\describe{
\item{\code{memory}}{The amount of memory used by the job}

\item{\code{cpus}}{The number of cpus used by the job}

\item{\code{walltime}}{The maximum amount of time the job should run}

\item{\code{qos}}{Name of the "quality of service" (qos) that the job should be submitted as}

\item{\code{output.file}}{The path to the Slurm output file}

\item{\code{email.complete}}{If used, Slurm will email the given email address when the job is finished.}

\item{\code{other.args}}{A string containing additional flags and values to pass to Slurm. These will overwrite previously given parameters when submitted.}
}}

\seealso{
\code{\linkS4class{Submission}}, \code{\link{submit}}
}
