% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Submission-class}
\alias{Submission-class}
\alias{Submission}
\title{Submission constructor}
\description{
An S4 class used to create a job submission.
}
\section{Slots}{

\describe{
\item{\code{script.file}}{Relative path to the script}

\item{\code{cluster}}{Address of cluster you want to submit to.}

\item{\code{profile}}{A \linkS4class{JobProfile} type object.}

\item{\code{script.args}}{String containing all arguments to the submitted Rscript to be executed remotely on the cluster.}

\item{\code{conda}}{String containing the name of a valid conda environment on the cluster to be used to execute R.}

\item{\code{singularity}}{String path or URI of a Singularity container to execute Rscript inside of.}

\item{\code{r.bin.args}}{String containing all arguments to pass to the \code{Rscript} binary.}

\item{\code{r.data}}{String containing the name of an R data object in the search path.}

\item{\code{local.data}}{String containing a list of paths to one or more local file(s). WARNING: Large files may take a long time to sync.}

\item{\code{target.dir}}{Path to directory to create .slurmsubmit/ directory.}
}}

\seealso{
\code{\linkS4class{JobProfile}}, \code{\link{submit}}
}
