% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaxsubmit.R
\name{submit}
\alias{submit}
\title{Submit a new Rscript to the cluster}
\usage{
submit(
  script = "",
  cluster = "sumner",
  slurm.args = "-q batch",
  rscript.args = "",
  conda = "",
  container = "docker://rocker/r-base:latest",
  r.bin.args = "",
  target.dir = "~"
)
}
\arguments{
\item{script}{Relative path to the script}

\item{cluster}{Name of cluster you want to submit to (sumner or winter).}

\item{slurm.args}{String containing all \code{sbatch} arguments for your submission.}

\item{rscript.args}{String containing all arguments to the submitted Rscript to be executed remotely on the cluster.}

\item{conda}{String containing the name of a valid conda environment on the cluster to be used to execute R.}

\item{container}{String path or URI of a Singularity container to execute Rscript inside of.}

\item{r.bin.args}{String containing all arguments to pass to the \code{Rscript} binary.}

\item{target.dir}{Path to tier1 directory to create .jaxsubmit/ directory.}
}
\value{
Slurm job ID of submitted script
}
\description{
This function is used to submit a new Rscript job to the cluster. It will fill in a
cookie-cutter Slurm submission script and execute R either inside a Singularity container
or with an existing conda environment.
}
\examples{
submit(script="example.R", slurm.args="-q batch -t 4:00:00 -o '.jaxsubmit/\%j.output'"
container="docker://rocker/tidyverse:latest", r.bin.args="--save")
}
